LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-relayout/multiple-aligned-targets/prefer-targeted-element-both-axes-expected.txt



FAIL scroller selects targeted area among multiple aligned areas on both axes assert_equals: scroller prefers #target over #a on X axis expected 200 but got 210
